# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Marketplace Nigeria API Documentation'
  description: 'Complete REST API documentation for Marketplace Nigeria platform. This API allows you to manage adverts, users, messages, payments, and more.'
  version: '3.0'
servers:
  -
    url: 'http://127.0.0.1:8030'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can obtain your authentication token by logging in through the `/api/login` endpoint. The token should be included in the Authorization header as a Bearer token for all authenticated requests.'
security:
  -
    default: []
paths:
...
  /api/messages:
    post:
      summary: ''
      operationId: postApiMessages
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                advert_id:
                  type: integer
                  description: 'The <code>id</code> of an existing record in the adverts table.'
                  example: 16
                receiver_id:
                  type: string
                  description: 'The <code>user_id</code> of an existing record in the users table.'
                  example: tempora
                message_content:
                  type: string
                  description: 'Must not be greater than 1000 characters.'
                  example: xn
                  nullable: true
                images:
                  type: array
                  description: 'Must be an image. Must not be greater than 12048 kilobytes.'
                  items:
                    type: string
                    format: binary
                    nullable: true
              required:
                - advert_id
                - receiver_id
      security: []
  '/api/messages/conversation/{advertId}/{receiverId}':
    get:
      summary: ''
      operationId: getApiMessagesConversationAdvertIdReceiverId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      security: []
    parameters:
      -
        in: path
        name: advertId
        description: ''
        example: aut
        required: true
        schema:
          type: string
      -
        in: path
        name: receiverId
        description: ''
        example: adipisci
        required: true
        schema:
          type: string
  '/api/messages/advert/{advertId}':
    get:
      summary: ''
      operationId: getApiMessagesAdvertAdvertId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      security: []
    parameters:
      -
        in: path
        name: advertId
        description: ''
        example: a
        required: true
        schema:
          type: string
...
  '/api/messages/{messageId}/read':
    put:
      summary: ''
      operationId: putApiMessagesMessageIdRead
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      security: []
    parameters:
      -
        in: path
        name: messageId
        description: ''
        example: ut
        required: true
        schema:
          type: string
  '/api/messages/conversation/{advertId}/{userId}/read':
    put:
      summary: ''
      operationId: putApiMessagesConversationAdvertIdUserIdRead
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      security: []
    parameters:
      -
        in: path
        name: advertId
        description: ''
        example: est
        required: true
        schema:
          type: string
      -
        in: path
        name: userId
        description: ''
        example: consequatur
        required: true
        schema:
          type: string
  /api/messages/archive:
    post:
      summary: ''
      operationId: postApiMessagesArchive
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advert_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the adverts table.'
                  example: et
                other_user_id:
                  type: string
                  description: 'The <code>user_id</code> of an existing record in the users table.'
                  example: est
              required:
                - advert_id
                - other_user_id
      security: []
  /api/messages/unarchive:
    post:
      summary: ''
      operationId: postApiMessagesUnarchive
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advert_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the adverts table.'
                  example: cumque
                other_user_id:
                  type: string
                  description: 'The <code>user_id</code> of an existing record in the users table.'
                  example: temporibus
              required:
                - advert_id
                - other_user_id
      security: []
...